Fix many (but not all) of the issues raised by -Wconversion
git-svn-id: http://root.cern.ch/svn/root/trunk@38402 27541ba8-7e3a-0410-8455-c3a389f83636

+++ cint/cint/inc/FastAllocString.h
@@ -52,7 +52,7 @@ G__EXPORT
 #endif
 G__FastAllocString {
 public:
-   G__FastAllocString(int reqsize = 1024): fBuf(0), fCapacity(reqsize) {
+   G__FastAllocString(size_t reqsize = 1024): fBuf(0), fCapacity(reqsize) {
       // GetBuf takes as parameter the size in bytes
       // and modify the parameter (fBucket) to hold the 
       // bucket number.
